# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AssetClass": {
        "description": "`asset_class` vocabulary (design §4.2 column comment).",
        "enum": [
          "equity",
          "bond",
          "fund",
          "option",
          "crypto",
          "commodity",
          "real_estate",
          "private",
          "other"
        ],
        "type": "string"
      },
...
      "Currency": {
        "description": "ISO 4217 currency code (`char(3)` in DDL), uppercase on the wire — `snake_case`\nwould mangle codes. Closed set for now; extend as regimes land (visible in the\nschema snapshot).",
        "enum": [
          "EUR",
          "GBP",
          "USD"
        ],
        "type": "string"
      },
      "DisclosureRecord": {
        "description": "One canonical disclosure record (Gold `disclosure_record`, design §4.2).\n`verification_state` is present on EVERY record — honesty travels with\nthe data.",
        "properties": {
          "as_of_date": {
            "description": "Snapshot date for holdings.",
            "format": "date",
            "type": [
              "string",
              "null"
            ]
          },
          "asset_class": {
            "$ref": "#/components/schemas/AssetClass",
            "description": "Asset class vocabulary."
          },
          "asset_description_raw": {
            "description": "Asset description exactly as filed, always kept (invariant 2).",
            "type": "string"
          },
          "created_at": {
            "description": "When the record row was created.",
            "format": "date-time",
            "type": "string"
          },
          "details": {
            "description": "Contract-typed regime payload (validated per (regime, `record_type`)\nJSON Schema at promotion, invariant 5).",
            "type": "object"
          },
          "event_date": {
            "description": "Best event date: `coalesce(transaction, notified, as_of)` (generated).",
            "format": "date",
            "type": [
              "string",
              "null"
            ]
          },
          "extracted_by": {
            "description": "Parser id / model+prompt version that produced this record.",
            "type": "string"
          },
          "extraction_confidence": {
            "description": "Extractor confidence in `[0, 1]`.",
            "format": "float",
            "type": [
              "number",
              "null"
            ]
          },
          "filing_id": {
            "description": "Source filing this record came from.",
            "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
            "type": "string"
          },
          "fingerprint": {
            "description": "Idempotency fingerprint (audit trail).",
            "type": "string"
          },
          "id": {
            "description": "Record ULID (time-sortable; the pagination cursor).",
            "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
            "type": "string"
          },
          "instrument_id": {
            "description": "Resolved instrument — `null` below threshold, never guessed (invariant 3).",
            "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
            "type": [
              "string",
              "null"
            ]
          },
          "notified_date": {
            "description": "Date the interest/change was notified.",
            "format": "date",
            "type": [
              "string",
              "null"
            ]
          },
          "owner": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Owner",
                "description": "Whose asset (self/spouse/...)."
              }
            ]
          },
          "politician_id": {
            "description": "The politician the record concerns.",
            "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
            "type": "string"
          },
          "record_type": {
            "$ref": "#/components/schemas/RecordType",
            "description": "One of the four observation types."
          },
          "regime_id": {
            "description": "Disclosure regime the record was filed under.",
            "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
            "type": "string"
          },
          "side": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Side",
                "description": "Transaction direction; present when `record_type == transaction`."
              }
            ]
          },
          "supersedes_record_id": {
            "description": "Superseded record in a correction chain (supersede, never update).",
            "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
            "type": [
              "string",
              "null"
            ]
          },
          "transaction_date": {
            "description": "Date the transaction happened.",
            "format": "date",
            "type": [
              "string",
              "null"
            ]
          },
          "value": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ValueInterval",
                "description": "Declared value band; bounds are decimal STRINGS (invariant 7)."
              }
            ]
          },
          "verification_state": {
            "$ref": "#/components/schemas/VerificationState",
            "description": "Two-stage publication state (design §7.1) — on every record."
          }
        },
        "required": [
          "id",
          "filing_id",
          "politician_id",
          "regime_id",
          "asset_description_raw",
          "record_type",
          "asset_class",
          "verification_state",
          "extracted_by",
          "fingerprint",
          "details",
          "created_at"
        ],
        "type": "object"
      },
      "ErrorBody": {
        "description": "The error envelope every non-2xx response carries.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail",
            "description": "The error itself."
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "ErrorDetail": {
        "description": "Machine-readable error code plus a human-readable message.",
        "properties": {
          "code": {
            "description": "Stable machine-readable code (e.g. `invalid_cursor`, `not_found`).",
            "examples": [
              "invalid_cursor"
            ],
            "type": "string"
          },
          "message": {
            "description": "Human-readable explanation.",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "FilingProvenance": {
        "description": "Filing provenance: the source filing the record came from.",
        "properties": {
          "external_id": {
            "description": "Source-native filing id when the source has one.",
            "type": [
              "string",
              "null"
            ]
          },
          "filed_date": {
            "description": "Date the filing was filed.",
            "format": "date",
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "description": "Filing ULID.",
            "type": "string"
          },
          "published_at": {
            "description": "When the government made it public.",
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id"
        ],
        "type": "object"
      },
...
      "InstrumentId": {
        "description": "ULID (26-char Crockford base32, time-sortable)",
        "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
        "type": "string"
      },
...
      "Owner": {
        "description": "`owner` CHECK: whose asset the record concerns. `self` is a Rust keyword,\nhence the `Self_` variant with an explicit wire rename.",
        "enum": [
          "self",
          "spouse",
          "dependent",
          "joint",
          "unknown"
        ],
        "type": "string"
      },
...
      "PoliticianId": {
        "description": "ULID (26-char Crockford base32, time-sortable)",
        "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
        "type": "string"
      },
...
      "Provenance": {
        "description": "Full provenance of one record: filing, archived raw document, and the\nregime it was filed under (with its methodology metadata).",
        "properties": {
          "filing": {
            "$ref": "#/components/schemas/FilingProvenance",
            "description": "The source filing."
          },
          "raw_document": {
            "$ref": "#/components/schemas/RawDocumentProvenance",
            "description": "Our archived copy of the official document."
          },
          "regime": {
            "$ref": "#/components/schemas/Regime",
            "description": "The disclosure regime the record was filed under."
          }
        },
        "required": [
          "filing",
          "raw_document",
          "regime"
        ],
        "type": "object"
      },
      "RawDocumentProvenance": {
        "description": "Raw-document provenance: our archived copy of the official document\n(invariant 2: raw is sacred — official link + archived copy, design §7.3).",
        "properties": {
          "fetched_at": {
            "description": "When we fetched the document.",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "Raw document ULID.",
            "type": "string"
          },
          "sha256": {
            "description": "sha256 of the archived Bronze bytes (integrity + citability).",
            "type": "string"
          },
          "source_url": {
            "description": "Official source URL the document was fetched from.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "sha256",
          "fetched_at"
        ],
        "type": "object"
      },
      "RecordDetail": {
        "description": "One record with its trust surface (design §7.3): provenance plus the full\nsupersession history in both directions (invariant 1: corrections insert\nsuperseding rows, so history is a chain of records).",
        "properties": {
          "provenance": {
            "$ref": "#/components/schemas/Provenance",
            "description": "Where it came from."
          },
          "record": {
            "$ref": "#/components/schemas/DisclosureRecord",
            "description": "The record itself."
          },
          "superseded_by": {
            "description": "Corrections that (transitively) supersede this one, ascending ULID.",
            "items": {
              "$ref": "#/components/schemas/DisclosureRecord"
            },
            "type": "array"
          },
          "supersedes": {
            "description": "Records this one (transitively) supersedes — the corrected history,\nascending ULID (= insertion-time) order.",
            "items": {
              "$ref": "#/components/schemas/DisclosureRecord"
            },
            "type": "array"
          }
        },
        "required": [
          "record",
          "provenance",
          "supersedes",
          "superseded_by"
        ],
        "type": "object"
      },
...
      "RecordPage": {
        "description": "One page of records; `next_cursor` is the last item's ULID, `null` when\nthe listing is exhausted (design §6.1: cursor pagination on ULIDs).",
        "properties": {
          "items": {
            "description": "Records in ascending ULID (= insertion-time) order.",
            "items": {
              "$ref": "#/components/schemas/DisclosureRecord"
            },
            "type": "array"
          },
          "next_cursor": {
            "description": "Pass back as `cursor` to fetch the page after this one; `null` at the end.",
            "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
...
      "RecordType": {
        "description": "`record_type` CHECK: the four observation types, one table (design D1).",
        "enum": [
          "transaction",
          "holding",
          "interest",
          "change_notification"
        ],
        "type": "string"
      },
      "Regime": {
        "description": "One disclosure regime — a jurisdiction body's disclosure rules and the\ntransparency-scorecard metadata (design §6.1/§7.3).",
        "properties": {
          "body": {
            "description": "Disclosing body, e.g. `US House`.",
            "type": "string"
          },
          "cadence": {
            "description": "Free-form filing cadence description.",
            "type": [
              "string",
              "null"
            ]
          },
          "disclosure_lag_days": {
            "description": "Statutory maximum disclosure lag in days.",
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          },
          "effective_from": {
            "description": "Date the regime's rules took effect.",
            "format": "date",
            "type": "string"
          },
          "effective_to": {
            "description": "Date the rules stopped applying; `null` while current.",
            "format": "date",
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "description": "Regime ULID (stable adapter constant).",
            "type": "string"
          },
          "jurisdiction_id": {
            "description": "Jurisdiction the regime belongs to (`jurisdiction.id`, e.g. `us`).",
            "type": "string"
          },
          "regime_type": {
            "description": "`transaction_report` | `periodic_declaration` | `change_notification`\n| `none`.",
            "type": "string"
          },
          "source_url": {
            "description": "Official source landing page.",
            "type": [
              "string",
              "null"
            ]
          },
          "value_precision": {
            "description": "Value precision the regime discloses: `exact` | `banded` |\n`categorical` | `none`.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "jurisdiction_id",
          "body",
          "regime_type",
          "value_precision",
          "effective_from"
        ],
        "type": "object"
      },
...
      "Side": {
        "description": "`side` CHECK: direction of a transaction.",
        "enum": [
          "buy",
          "sell",
          "exchange"
        ],
        "type": "string"
      },
...
      "ValueInterval": {
        "description": "A money interval `low ..= high` in `currency`. Exact figures are `low == high`;\n`high == None` is an open-ended threshold (UK \"over £70,000\"). Construction is\nthe only door: `high < low` never exists — in memory or on the wire, because\nthe manual `Deserialize` impl funnels back through [`ValueInterval::new`].",
        "properties": {
          "currency": {
            "$ref": "#/components/schemas/Currency"
          },
          "high": {
            "type": [
              "string",
              "null"
            ]
          },
          "low": {
            "type": "string"
          }
        },
        "required": [
          "low",
          "currency"
        ],
        "type": "object"
      },
      "VerificationState": {
        "description": "`verification_state` CHECK: two-stage publication states (design §4.2).",
        "enum": [
          "unverified",
          "verified",
          "corrected",
          "disputed"
        ],
        "type": "string"
      }
    }
  },
  "info": {
    "description": "Worldwide politician financial-disclosure tracking. Cursor pagination on ULIDs; consistent error envelope; verification_state on every record. Every successful GET carries a strong ETag (sha256 of the body); requests with a matching If-None-Match receive 304 Not Modified. Authentication: `Authorization: Bearer gfk_...` API keys (create via /v1/keys; hashes only at rest). Requests without a key are served at the free tier. Tiers control record freshness (free: filings we discovered less than 24 hours ago are not yet visible; pro/data: real time) and the daily request quota (counted per account; over quota is 429 quota_exceeded in the standard envelope). Anonymous traffic has a light per-IP rate backstop (429 rate_limited).",
    "title": "govfolio API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/records": {
      "get": {
        "description": "# Errors\n`400` on a malformed cursor, limit or filter; `500` on backend failure —\nall in the consistent error envelope.",
        "operationId": "list_records",
        "parameters": [
          {
            "description": "Pagination cursor: the `id` of the last record on the previous page\n(from `next_cursor`). Pages begin strictly after it.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
              "type": "string"
            }
          },
          {
            "description": "Page size, `1..=200`; defaults to 50.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int32",
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Only records filed under a regime of this jurisdiction\n(`jurisdiction.id`, e.g. `us`).",
            "in": "query",
            "name": "jurisdiction_id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only records of this type.",
            "in": "query",
            "name": "record_type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RecordType"
            }
          },
          {
            "description": "Only records of this asset class.",
            "in": "query",
            "name": "asset_class",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AssetClass"
            }
          },
          {
            "description": "Only records concerning this politician.",
            "in": "query",
            "name": "politician_id",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PoliticianId"
            }
          },
          {
            "description": "Only records resolved to this instrument (below-threshold matches are\nNULL and never match — invariant 3).",
            "in": "query",
            "name": "instrument_id",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/InstrumentId"
            }
          },
          {
            "description": "Only records with `event_date` on or after this date.",
            "in": "query",
            "name": "event_date_from",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "description": "Only records with `event_date` on or before this date.",
            "in": "query",
            "name": "event_date_to",
            "required": false,
            "schema": {
              "format": "date",
              "type": "string"
            }
          },
          {
            "description": "Only records whose declared value band can reach this amount or more\n(decimal string — invariant 7).",
            "in": "query",
            "name": "value_min",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only records whose declared value band can stay at this amount or less\n(decimal string — invariant 7).",
            "in": "query",
            "name": "value_max",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only records in this verification state.",
            "in": "query",
            "name": "verification_state",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/VerificationState"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordPage"
                }
              }
            },
            "description": "One page of disclosure records; every record carries verification_state"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Malformed cursor, limit or filter"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Lists disclosure records in ascending ULID order.",
        "tags": [
          "records"
        ]
      }
    },
    "/v1/records/{id}": {
      "get": {
        "description": "# Errors\n`404` for an unknown record; `500` on backend failure — all in the\nconsistent error envelope.",
        "operationId": "get_record",
        "parameters": [
          {
            "description": "Record ULID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordDetail"
                }
              }
            },
            "description": "The record with provenance and supersession history"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unknown record"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "summary": "Fetches one record with provenance and its supersession chain.",
        "tags": [
          "records"
        ]
      }
    },
...
  },
...
}